requirements.txt
google-adk
dotenv
replicate
elevenlabs
moviepy